+++ server.js
@@ -1,0 +1,20 @@
+const jsonServer = require("json-server");
+const server = jsonServer.create();
+const router = jsonServer.router("./src/utils/api/data.json");
+const middlewares = jsonServer.defaults({
+  static: "./build",
+});
+const port = process.env.PORT || 4000;
+
+server.use(middlewares);
+server.use(
+  jsonServer.rewriter({
+    "/api/*": "/$1",
+  })
+);
+
+server.use(router);
+
+server.listen(port, () => {
+  console.log("server is running");
+});
